requirements.txt
streamlit==1.35.0
streamlit-option-menu==0.3.6
scikit-learn==1.4.2
matplotlib==3.8.4
inference==0.9.1
Pillow==10.3.0
google.generativeai